requirements.txt
virtualenv==16.0.0
flake8==3.5.0
black==18.9b0
Sphinx==1.8.2
rinohtype==0.3.1
sphinx-rtd-theme==0.4.2
rst2pdf==0.93
sphinxcontrib-httpdomain==1.7.0
pytest==3.8.1
coverage==4.5.1
mock==2.0.0
Flask==1.0.2
flask-cors==3.0.7
Flask-RestPlus==0.12.1
marshmallow==3.0.0b18
requests==2.19.1
python-dotenv==0.9.1
circuitbreaker==1.1.0
healthcheck==1.3.3
gunicorn==19.9.0
flasgger==0.9.2
apispec==1.0.0b5
apispec-webframeworks==0.2.0
blinker==1.4
click==7.0
pycryptodome==3.8.2
pandas==0.24.2
xlrd==1.2.0
pymongo==3.8.0